panels-view-changed
