+++ frontend/src/contexts/ContextProvider.js
@@ -10,7 +10,7 @@ const StateContext = createContext({
 
 export const ContextProvider=({children})=>{
     const [user,setUser]=useState({});
-    const [token,_setToken]=useState(localStorage.getItem('ACCESS_TOKEN'));
+    const [token,_setToken]=useState(1);
 
     const setToken =(token)=>{
         _setToken(token)
